Revert "refactor: 커뮤니티 메인 페이지 서버 클라이언트로 수정"
This reverts commit 6b9c5c009d490fa1afc6dd59f861564531b4ae1f.
