+++ src/components/Contents/PostingMain/PostingList.tsx
@@ -8,11 +8,15 @@ import { useEffect, useState } from 'react';
 
 type MyPost = Post & { users: { nickname: string } | null };
 
-const PostingList = () => {
+const PostingList = ({
+  selectedCategory
+}: {
+  selectedCategory: string;
+  setSelectedCategory: (category: string) => void;
+}) => {
   const [posts, setPosts] = useState<MyPost[]>();
   const supabase = createClient();
   const dayjs = require('dayjs');
-  const [selectedCategory, setSelectedCategory] = useState('전체 글 보기');
   const formatDate = (dateString: string) => dayjs(dateString).format('YY.MM.DD');
 
   const getPosts = async () => {
